\name{rioja-internal}
\alias{.max.bias}
\alias{.r2}
\alias{.RMSE}
\alias{.crossval}
\alias{.predict}
\alias{.check.data}
\alias{.print.crossval}
\alias{.performance}
\alias{.print.performance}
\alias{.set.rand.seed}
\alias{.get.rand}
\alias{predict.internal.WA}
\alias{predict.internal.WAPLS}
\alias{predict.internal.IKFA}
\alias{predict.internal.MLRC}

\title{Internal rioja functions. }
\description{
Internal rioja functions.
}
\usage{
.max.bias(resi, x, ng = 10)
.rmse(x)
.r2(x, obs)
}

\details{
  Internal \code{rioja} functions, not to be called by the user.
}

\author{ 
Steve Juggins
}

\keyword{ internal }
