\name{rioja-package}
\alias{rioja-package}
\alias{rioja}
\docType{package}
\title{
Analysis of Quaternary Science Data
}
\description{
\code{rioja}: An R package for the analysis of Quaternary science data. Contains functions for constrained clustering, transfer functions, and plotting stratigraphic data.
}

\details{
The \emph{rioja} package contains a number of tools for analysing and visualising (bio)stratigraphic data and for developing palaeoecological transfer functions from a dataset of modern species counts and environmental measurements. Resulting models can be cross-validated using the \emph{crossval} function, which allows internal cross-validation using leave-one-out, leave-n-out, bootstrapping or h-block cross-validation.

\packageIndices{rioja}
}

\author{
Steve Juggins

Maintainer: Steve Juggins <Stephen.Juggins@ncl.ac.uk>
}
\keyword{ package }

